refactor:#32 validate 함수 파일 위치 이동 및 editProfile 스키마 추가
