+++ src/lib/utils/validation/auth-validate.ts
@@ -24,3 +24,12 @@ export const signupSchema = z
   });
 
 export type SignupFormData = z.infer<typeof signupSchema>;
+
+export const editProfileSchema = z.object({
+  nickname: z
+    .string({ required_error: '닉네임을 입력해주세요.' })
+    .min(2, '닉네임은 최소 2자 이상이어야 합니다.')
+    .max(10, '닉네임은 10자 이하여야 합니다.')
+});
+
+export type editProfileData = z.infer<typeof editProfileSchema>;
